// src/types/layer-types.ts
export interface XYZLayerProps {
  title: string;
  maxZoom: number;
  attributions?: string;
  attributionsCollapsible?: boolean;
  url: string;
  image?: string;
}

export interface BaseMapProps {
  title: string;
  maxZoom: number;
  attributions?: string;
  attributionsCollapsible?: boolean;
  url: string;
  image?: string;
}